Add files via upload
In this Netflix data analysis project, I uncovered key insights by leveraging Python's pandas and matplotlib libraries. I discerned the dominance of either movies or TV shows in the dataset, pinpointed the year with the lowest releases, tallied distinct ratings, and generated a frequency table showcasing the prevalence of content categories. This exploration offers a snapshot of Netflix's content landscape, shedding light on trends, viewer preferences, and the distribution of entertainment genres.